Good News! Work starting again
I have plans to start working on this framework again. However due to exams and other difficulties I will not spend too much time in this but enough to keep it going. And as always, one person is not enough for a framework like this so if anybody is willing to help, please either contact me or pick something from the README and create a pull request.